+++ lib/widgets/trip_viewer_page.dart
@@ -282,7 +282,8 @@ class _TripViewerPageState extends State<TripViewerPage> {
         0.000025; // Approx 2.7 meters, adjust as needed
 
     do {
-      double effectiveDisplacementOrder = displacementRank + attempt as double;
+      double effectiveDisplacementOrder =
+          displacementRank.toDouble() + attempt.toDouble();
       double angle =
           angularSeed +
           effectiveDisplacementOrder *
